requirements.txt
# Core data analysis libraries
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# BigQuery and Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-auth==2.23.4

# Jupyter and notebooks
jupyter==1.0.0
ipykernel==6.25.2

# Data processing
pyarrow==14.0.2
fastparquet==2023.10.1

# Statistical analysis
scipy==1.11.4
scikit-learn==1.3.2

# Visualization
kaleido==0.2.1
dash==2.14.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
